requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-phonenumber-field==7.0.0
phonenumbers==8.13.1
sqlparse==0.4.3
